Sketch the assembler interface
Most part of the parse tree is included in this commit.
Most of the instructions are still missing, the whole data management is still missing.
The Instruction* classess have been refactored. In particular, most of the logic to
manage operands, and assemble instructions is moved there. This logic is still to be
implemented.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>