Management of microoperations
Instruction classes from 0 to 6 have been modified so as to introduce as
well the generation of microoperations during the assembly phase.

NOTE: This is a merge of contributed code. It is the result of several
squashed commits.
Code has been reorganized to meet the code standards of the project and
to uniform the syntax as much as possible. The semantics of the operations
have not been touched, though.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>