Redesign the Syntax Error Highlighter
This works now in a separate runnable, which avoids a certain
number of runtime errors if the taks takes too long.
Nevertheless, to do this some interaction with the Parsing API
has been changed, and in fact the ParserResultTask's run()
method is never called. Still have to investigate on this.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>