Move Program packages
Program packages are moved to the simulator. This is done in order
to avoid a cyclic dependency on the modules.
This change has required to decouple ParseException from Program
methods. Therefore, a new ProgramException is included. This exception
is then catched by the assembler, which in turn fires back a
ParseException, in order for the assemblying action to catch it
and display errors.
ProgramExceptions now keep in the text the line in the source code
at which the error was generated.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>