Assembler and Disassembler of instructions
This commit sketches the assembler of the instructions in a much better
way than before. All opcodes are now correctly assembled. There is still
the remainder of the bytes of the instructions to be assembled, this regards
mostly Class 1 and Class 2.
Additionally, all instruction and operand classes implement toString()
methods which can be used to disassemble them.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>