Lazy relocation of labels
When a label is encountered in an instruction, it's value is not retrieved
immediately. Rather, relocation entries are generated.
At the end of the compiling process, the list of relocation entries is scanned,
and all relocations (depending on labels' values) are applied to operands.
This allows to account for scenarios where an instruction is targeting a
label which has not yet been encountered (for example, jumps to instructions
which appear later in a program).

This has implications on the readability of error messages (it's not possible to
know now which instruction was using which label).

Relocation does not take into account yet binary encoding of the instructions.

Signed-off-by: Alessandro Pellegrini <pellegrini@dis.uniroma1.it>