Configuration Dialog
Introducing the configuration dialog, accessible from the menu.
If changes are made to the configuration, they are either dispatched
(two are still missing, due to a required change to the dispatcher)
or the user is informed of the need for a restart.

Configuration is also stored in the configuration file on disk.

Signed-off-by: Alessandro Pellegrini <a.pellegrini@ing.uniroma2.it>