Properly handle Dock icon on macOS
The JAR icon is not automatically rendered in the dock bar on macOS, so
it is required to programmatically change it.

In order to do that, to avoid accessing native stuff, I have jumped to
Java 11 as minimum dependency, because the class to access that facility
in a java-way was introduced in Java 9. Not too bad, JRE 8u321 is going
to be discontinued soon, it was just a matter of affection.

Signed-off-by: Alessandro Pellegrini <a.pellegrini@ing.uniroma2.it>