[WIP] Reworking relocations and program structure
The current program structure is neat and interesting, but too hard to
maintain and complete, given the short amount of time to finish an early
version of this project.

I am therefore rolling back to a worse implementation based on linked
data structures.
Text and data are currently reworked (but untested), and the stack is
therefore gone.

Still have to implement proper relocations. They shouldn't be a pain if
I use an extra element in an instruction allowing to point to something
in memory. Should only see if I have a proper super class for
instructions and data.

Signed-off-by: Alessandro Pellegrini <a.pellegrini@ing.uniroma2.it>