Implemented FLAGS update logic
The algorithm(s) to update flags are introduced. There is a general
algorithm in SimulationController that updates flags as if they were
the results of an addition. This is fine for most instructions, as
also the hardware implementation does so.

Then, special cases for other instructions (e.g., shifts) are handled
in the actual instruction algorithms.

Signed-off-by: Alessandro Pellegrini <a.pellegrini@ing.uniroma2.it>